Cleanup of Orleans logging layer.
TraceLogger has been broken into three classes and deprecated. It is still maintained mainly to support programmatic log configuration until customers can migrate to the new pattern.
 LoggerImpl replaces TraceLoggers logger implementation role. It is internal and all public api surfaces now use Logger.
 LogManager replaces TraceLoggers support of programmatic logging settings.
 LogFormatter replaces TraceLoggers log formatting utility functions.

+++ src/OrleansRuntime/Streams/PubSub/PubSubRendezvousGrain.cs
@@ -222,7 +222,7 @@ namespace Orleans.Streams
 
         private void RemoveProducer(PubSubPublisherState producer)
         {
-            logger.Warn((int)ErrorCode.Stream_ProducerIsDead,
+            logger.Warn(ErrorCode.Stream_ProducerIsDead,
                 "Producer {0} on stream {1} is no longer active - permanently removing producer.",
                 producer, producer.Stream);
